remove simple crowdfunding contract, rename CrowdfundingExt to Crowdfunding
